requirements.txt
Django
djangorestframework
Pillow
qrcode
